LPS-146974 Move to proper package.
