+++ modules/apps/layout/layout-taglib/src/main/java/com/liferay/layout/taglib/internal/servlet/ServletContextUtil.java
@@ -26,8 +26,8 @@ import com.liferay.layout.adaptive.media.LayoutAdaptiveMediaProcessor;
 import com.liferay.layout.display.page.LayoutDisplayPageProviderTracker;
 import com.liferay.layout.list.retriever.LayoutListRetrieverTracker;
 import com.liferay.layout.list.retriever.ListObjectReferenceFactoryTracker;
+import com.liferay.layout.util.CollectionPaginationHelper;
 import com.liferay.layout.util.LayoutClassedModelUsageRecorder;
-import com.liferay.layout.util.structure.CollectionPaginationHelper;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.segments.SegmentsEntryRetriever;
